refactor: split docs markdown metadata helpers
