+++ crates/ox_content_docs/src/markdown/labels.rs
@@ -1,0 +1,44 @@
+use crate::string_builder::StringBuilder;
+
+pub(super) fn normalize_signature(signature: Option<&str>) -> Option<String> {
+    let signature = signature?;
+    let normalized = signature.split_whitespace().collect::<Vec<_>>().join(" ");
+    let mut value = normalized.as_str();
+
+    for prefix in [
+        "export ",
+        "declare ",
+        "abstract ",
+        "async function ",
+        "function ",
+        "class ",
+        "interface ",
+        "type ",
+    ] {
+        if let Some(stripped) = value.strip_prefix(prefix) {
+            value = stripped;
+        }
+    }
+
+    Some(value.trim().to_string()).filter(|value| !value.is_empty())
+}
+
+pub(super) fn format_kind_label(kind: &str) -> &str {
+    match kind {
+        "function" => "fn",
+        "interface" => "interface",
+        "class" => "class",
+        "type" => "type",
+        "const" => "const",
+        _ => kind,
+    }
+}
+
+pub(super) fn format_count_label(count: usize, singular: &str, plural: Option<&str>) -> String {
+    let label = if count == 1 { singular } else { plural.unwrap_or(singular) };
+    let mut out = StringBuilder::new();
+    out.push_usize(count);
+    out.push_char(' ');
+    out.push_str(label);
+    out.into_string()
+}
